examples/jwt/requirements.txt
pycrypto==2.6.1
python-jwt==2.0.1
